Do not rely on global constructors.
This makes sure that no global constructors are required to run in the main
MUNT code. Global constructors have an implementation specific execution
order/time. There are only some minimal guarantees by the C++ standard.

In ScummVM these can be an issue for some obscure platforms because they might
not even get executed at all. This is not standard conforming but the reason
we disallow code with global constructors.

The change is not exactly pretty since it removes the visibility control of
ControlROMFeatureSet. I am not sure whether this data type is going to be
extended in the future or not. If it isn't, it might be possible to just
replace it directly with a boolean value.

clang's -Wglobal-constructors warning option allows to check whether global
constructors are used.