Introduced "nice amp ramp" mode
In this mode, we want to ensure that amp ramp never jumps to the target
value and always gradually increases or decreases. It seems that real units
do not bother to always check if a newly started ramp leads to an amp jump.
We also prefer the quality improvement over the emulation accuracy,
so this mode is enabled by default.

WIP. We only fix the sustain phase atm.
Still, other phases need to improve.