Optimised "nice amp ramp" mode towards slower transitions
Restored accurate calculation of target amp
when a ramp is started during sustain.
We now only ensure the correct ramp direction,
so it doesn't jump. This results in slower
amp transitions (because TVA is no longer aware
of actual current amp value), though. However,
the resulting sounding is even more pleasant
due to absence of fast amp changes which were
previously observed when the attack phase is over.