Build: added simplified support for precompiled headers
Enabling precompiled headers support significantly decreases
compilation time of mt32emu-qt. It was already simple
to enable in Visual Studio projects unlike the other build systems.
Now, configuration of precompiled headers can be easily done
when working, for example, with QtCreator. Supported compilers are
MSVC, GNU C++ and Clang.

To enable precompiled headers support, the pathname of the header
to be used as precompiled is set in mt32emu-qt_PRECOMPILED_HEADER property.