Added full support for Qt 4.5.3 in MasterClock
For Qt versions lower than 4.7.0, we now use the system clock
as a fall-back. We could limit the lower supported Qt version to 4.8.0
on platforms other than Windows but this wouldn't be a big win
because QElapsedTimer also falls back to the system clock
when there is no supported monotonic clock found. Our support
for POSIX clock should be further improved though.