Build: For mt32emu, changed semantic of the build option libmt32emu_WITH_INTERNAL_RESAMPLER
When disabled, the built-in resampler support is never included in the build
regardless of whether an external resampler library presence.
In case no external resampler library is found, a simple SRC stub
is used that outputs at the internal synth sample rate (or is muted).
This is mainly intended to reduce the library size when the sample rate
conversion is handled externally.