Fixed a bug in libsmf with parsing long meta events
When determining the length of a meta event,
the format was assumed to be

    0xFF 0xwhatever 0xlength and then "length" bytes.

However, the standard MIDI file's spec defines a variable-length
field to contain the event length.
That caused meta events with payload length above 130 bytes
to be parsed incorrectly. This is weird because the code elsewhere
handles them properly.