Introduced realtime rendering mode in QSynth
When JACKAudioStream is started with an exclusive MIDI session,
it now enables the realtime rendering mode in QSynth. Specifically,
the renderer no longer blocks the calling thread and handles external
changes of the synth settings on a best effort basis. The synth state
updates are now offloaded to a separate thread and disseminated from there.