Fixes in mt32emu_win32drv_setup
It no longer fails to match a device registered in the PnP manager with
a hardware ID that differs in case only. Apparently, the hdwwiz may add
the device with the hardware ID in all lower case. It may also not register
the hardware ID at all, so added a fallback to matching via the device
description as well.