Fix effc++ warnings in libmt32emu interface
When users of the library compile code with -Weffc++, this class
triggers warning:

  ‘class MT32Emu::Service’ has pointer data members
  but does not declare ‘MT32Emu::Service(const MT32Emu::Service&)’
  or ‘operator=(const MT32Emu::Service&)’

Class Service aggregates an owning raw pointer (mt32emu_context), which
is being taken over via constructor and freed in destructor - this might
lead user to accidentally double free the pointed object if a Service
object was accidentally copied through default copy c-tor or default
assignment operator.

Declare copy c-tor and operator= private to prevent creation of
accidental copy (this fixes the warning).

When upgrading this code to C++11, these declarations should be
marked with '= delete' (and won't need to be private any more).