mt32emu-qt now features floating window with emulated display (#50)
The window represents the emulated LCD and the MIDI MESSAGE LED only.
It is implicitly configured to be always-on-top, can be freely moved and
resized. Also, its opacity can be adjusted, if supported by the platform.
By default, the new window shows up when the main window is hidden. This
behaviour can be altered in the main menu / Options / Floating Display.