In mt32emu, improved the volume override function (#85)
When 0 volume is set on a part, it now gets completely muted. All
the currently played notes are released and new NoteOn messages are ignored
on that part. This is important since the real devices (and the emulation)
do not provide a way to silence a playing part completely - it may still
produce sound at a very low level even having the part volume set to 0.
Besides, avoiding generation of samples we don't want to be played saves
some CPU resources.