mt32emu: Added two new functions for retrieving sound group data (#91)
These provide a way to get names of all the standard timbres and how they
are assigned to the sound groups in the control ROM. Additionally, this
makes it easier to find the names of custom timbres.

Also, bumped target version 2.7.0 to enable detection of availability
of the new functions.