mt32emu-qt: Implemented UI for selecting timbres on each part (#70)
Added possibility to view timbre banks and select a desired instrument
on each voice part. The timbres are shown in a popup menu split by their
"sound groups" which is displayed when the user clicks on the instrument
name set on the related part in the Channel Status pane.

Also, bumped the minimum required version of mt32emu 2.7 and the target
version 1.11.0 of the mt32emu-qt app.