mt32emu-qt: Fixed some problems with Floating Display on multi-screen systems (#94)
When the application is restarted, the position of the floating window
should now be restored on the same screen where it was previously. Also,
the window geometry is now adjusted to fit into the screen when the display
it is show on is switched off, and each time the floating window shows up.