mt32emu-qt: Fixed occasional crashes when JACK audio device started (#106)
On some platforms, when a relatively small JACK buffer size is configured,
the JACK server manages to invoke the first audio rendering callback early
enough, so that the previous implementation of JACKClient sometimes hasn't
had initialised yet. The new implementation activates the JACK client in
a separate step, when the initialisation is complete.