mt32emu-qt: Improved output gain spin boxes in Synth properties (#104)
The spin boxes no longer receive value updates when the user is entering a
number directly. For the new value to take effect, the user has to move
the focus or press the Enter key (which also saves the change in settings).
This new behaviour makes the spin boxes a bit less convenient for
"spinning" (using the arrow keys or the mouse scroll) but obviously less
awkward for direct numerical input.